server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Jiho Park <jihopark7777@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = {extras = ["all"], version = "^0.87.0"}
uvicorn = "^0.20.0"
python-dotenv = "^0.21.0"

[tool.poetry.scripts]
start = "server.__init__:start_uvicorn"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
